requirements.txt
attrs==25.3.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
types-jsonschema==4.23.0.20241208
types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2
